+++ mcww_web/js/galleryButtons.js
@@ -26,7 +26,9 @@ function fixClipboardPaste(updatedElements) {
         container.dataset.clipboardFixAttached = "true";
     });
 
-    const galleries = updatedElements.querySelectorAll('.gallery-container:not(:has(button.paste))');
+    const galleries = updatedElements.querySelectorAll('.gallery-container:not(:has(.icon-button-wrapper button.paste)), ' +
+            '.upload-gallery:not(:has(.icon-button-wrapper button.paste))'
+    );
     galleries.forEach((container) => {
         if (container.parentElement.classList.contains("no-paste")) {
             return;