+++ Changelog.md
@@ -4,6 +4,7 @@
 - Mark canceled tasks in queue ui as blue, not as an error
 - Added "Download" button in text outputs
 - Removed never implemented "batch from directory". I think it's not possible in ComfyUI without crazy crutches
+- Support batch for audio inputs
 
 ### 2.2 – Improved files open and LLMs support
 - Support boolean primitive nodes (e.g. useful for "thinking" in text generation node)